backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
yfinance==0.2.24
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
python-dotenv==1.0.0
httpx==0.28.1

# Database dependencies
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Testing with database
pytest-postgresql==5.0.0
testcontainers==4.5.1

# Authentication (for future OAuth)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0